Add support for decompress chunk to work with deletes
This works now:
explain verbose
delete from foo where a = 1;

This crashes:
delete from foo where a = 1;
The correct ctid might be passed back.
But need to map from chunk to compressed chunk for
the delete.

Also need a way to idnicate which rows are to be rewritten back to the
db. (delete is actually an update in many cases).